requirements.txt
python-frontmatter
jinja2
boto3
